Statically link local NineAnimator Modules
This is required for NineAnimator as the current build script removes all .dylib frameworks, as a method of removing unnecessary system libraries from the app.
However, this would also remove the local dependancies as they were also dynamically linked, resulting in broken builds.

The remote NineAnimatorCommon package also needs to be updated to fully solve this issue.
This is being tracked in SuperMarcus/NineAnimatorCommon#1